完善了路径推荐接口。

+++ src/main/java/com/qg/www/models/Steps.java
@@ -32,6 +32,6 @@ public class Steps {
     /**
      * 该路段花费的时间；
      */
-    private Long time;
+    private Integer time;
 
 }
